First attempt at fixing #1
Move the IMAP connection within the loop logic so that each fetch
opens and closes a new connection. This may have been the issue because
imaplib is not async compatible.

While changing the loop's flow, I also modified the internal for loop
that fetches the numbers by removing the if statement checking if the
message number lists is non-empty. The return value of the call to
.split() returns a list anyways, so if the list of message numbers is
empty, the for loop effectively skips to the call to async sleep.

Also, some needless trailing whitespace was removed.